Added AutonomousDefinitions and TFLiteHandler (again)
hopefully everything will work this time. Hopefully the problem with TFLiteHandler was just the timer not being done properly
